ref: organize args 2/n (#3442)
* ref: organize args 2/n

* ref: organize args 2/n
